Update pending procedures

+++ pending/dms/update_analysis_job_processor_group_associations.sql
@@ -17,12 +17,14 @@ AS $$
 **      Sets jobs in the job list to be associated with the given analysis job processor group
 **
 **  Arguments:
-**    _newValue   ignore for now, may need in future
-**    _mode       'add', 'replace', 'remove'
+**    _jobList              Comma separated list of job numbers
+**    _processorGroupID     Processor group ID
+**    _newValue             Ignore for now, may need in future
+**    _mode                 'add', 'replace', 'remove'
 **
 **  Auth:   grk
 **  Date:   02/15/2007 Ticket #386
-**          02/20/2007 grk - fixed references to "Group" column in associations table
+**          02/20/2007 grk - Fixed references to "Group" column in associations table
 **                         - 'add' mode now removes association with any other groups
 **          03/28/2008 mem - Added optional parameter _callingUser; if provided, will populate field Entered_By with this name
 **          09/02/2011 mem - Now calling PostUsageLogEntry
@@ -58,8 +60,8 @@ BEGIN
     ---------------------------------------------------
     -- Resolve processor group ID
     ---------------------------------------------------
-    _gid := CAST(_processorGroupID as int);
-    --
+    _gid := try_cast(_processorGroupID, int);
+
 /*
     SELECT group_id INTO _gid
     FROM t_analysis_job_processor_group
@@ -77,7 +79,7 @@ BEGIN
 
     CREATE TEMP TABLE Tmp_Jobs (
         Job int
-    )
+    );
 
     ---------------------------------------------------
     -- Populate table from job list
@@ -85,50 +87,45 @@ BEGIN
 
     INSERT INTO Tmp_JobList (Job)
     SELECT DISTINCT Item
-    FROM public.parse_delimited_list(_jobList)
+    FROM public.parse_delimited_list(_jobList);
 
     ---------------------------------------------------
     -- Verify that all jobs exist
     ---------------------------------------------------
     --
-    _list := '';
-    --
-    SELECT
-        _list = _list + CASE
-        WHEN _list = '' THEN cast(Job as text)
-        ELSE ', ' || cast(Job as text)
-        END
+    SELECT string_agg(Job::text, ', ' ORDER BY Job)
+    INTO _list;
     FROM Tmp_JobList
-    WHERE
-        NOT job IN (SELECT job FROM t_analysis_job)
+    WHERE NOT job IN (SELECT job FROM t_analysis_job);
 
     If _list <> '' Then
-        _message := 'The following jobs were not in the database: "' || _list || '"';
+        _message := 'The following jobs were not in the database: ' || _list;
         _returnCode := 'U5202';
+
+        DROP TABLE Tmp_JobList;
         RETURN;
     End If;
 
-    _jobCount := 0;
+    SELECT COUNT(*)
+    INTO _jobCount
+    FROM Tmp_JobList;
 
-    SELECT COUNT(*) INTO _jobCount
-    FROM Tmp_JobList
-
-    _message := 'Number of affected jobs: ' || cast(_jobCount as text);
+    _message := format('Number of affected jobs: %s', _jobCount);
 
     ---------------------------------------------------
     -- Get rid of existing associations if we are
     -- replacing them with jobs in list
     ---------------------------------------------------
     --
-    If _mode = 'replace' Then
+    If _mode = 'replace' And _gid > 0 Then
         DELETE FROM t_analysis_job_processor_group_associations
-        WHERE (group_id = _gid);
+        WHERE group_id = _gid;
     End If;
 
     ---------------------------------------------------
     -- Remove selected jobs from associations
     ---------------------------------------------------
-    If _mode = 'remove' or _mode = 'add' Then
+    If _mode IN ('remove', 'add') Then
         DELETE FROM t_analysis_job_processor_group_associations
         WHERE job IN (SELECT job FROM Tmp_JobList);
             -- AND Group_ID = _gid  -- will need this in future if multiple associations allowed per job
@@ -139,31 +136,31 @@ BEGIN
     -- Add associations for new jobs to list
     ---------------------------------------------------
     --
-    If _mode = 'replace' or _mode = 'add' Then
-        INSERT INTO t_analysis_job_processor_group_associations
-            (job, group_id)
+    If _mode IN ('replace', 'add') Then
+        INSERT INTO t_analysis_job_processor_group_associations (job, group_id)
         SELECT job, _gid
-        FROM Tmp_JobList
+        FROM Tmp_JobList;
 
         _alterEnteredByRequired := true;
     End If;
 
     -- If _callingUser is defined, update entered_by in t_analysis_job_processor_group_associations
-    If char_length(_callingUser) > 0 And _alterEnteredByRequired Then
+    If char_length(Coalesce(_callingUser, '')) > 0 And _alterEnteredByRequired Then
         -- Call public.alter_entered_by_user for each processor job in Tmp_JobList
 
         CREATE TEMP TABLE Tmp_ID_Update_List (
             TargetID int NOT NULL
-        )
+        );
 
         CREATE INDEX IX_Tmp_ID_Update_List ON Tmp_ID_Update_List (TargetID);
 
         INSERT INTO Tmp_ID_Update_List (TargetID)
         SELECT Job
-        FROM Tmp_JobList
+        FROM Tmp_JobList;
 
         Call alter_entered_by_user_multi_id ('t_analysis_job_processor_group_associations', 'job', _callingUser);
 
+        DROP TABLE Tmp_ID_Update_List;
     End If;
 
     ---------------------------------------------------
@@ -174,7 +171,6 @@ BEGIN
     Call post_usage_log_entry ('Update_Analysis_Job_Processor_Group_Associations', _usageMessage);
 
     DROP TABLE Tmp_JobList;
-    DROP TABLE Tmp_ID_Update_List;
 END
 $$;
